+++ cmd/playpen/status.go
@@ -1,0 +1,42 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+	"strings"
+	"unicode"
+)
+
+// TrimRightSpace returns a slice of the string s, with all trailing white space
+// removed, as defined by Unicode.
+func TrimRightSpace(s string) string {
+	return strings.TrimRightFunc(s, unicode.IsSpace)
+}
+
+// Status reports the current status of the daemon
+func (d *DaemonService) Status(_ *http.Request, _ *EmptyArgs, reply *StringReply) error {
+	res := new(strings.Builder)
+	defer func() { reply.Message = TrimRightSpace(res.String()) }()
+
+	if !d.network.IsOkay() {
+		fmt.Fprintln(res, "Network overrides NOT established")
+	}
+	if d.cluster == nil {
+		fmt.Fprintln(res, "Not connected")
+		return nil
+	}
+	if d.cluster.IsOkay() {
+		fmt.Fprintln(res, "Connected")
+	} else {
+		fmt.Fprintln(res, "Attempting to reconnect...")
+	}
+	fmt.Fprintf(res, "  Context:       %s (%s)\n", d.cluster.Context(), d.cluster.Server())
+	if d.bridge != nil && d.bridge.IsOkay() {
+		fmt.Fprintln(res, "  Proxy:         ON (networking to the cluster is enabled)")
+	} else {
+		fmt.Fprintln(res, "  Proxy:         OFF (attempting to connect...)")
+	}
+	fmt.Fprintf(res, "  Interceptable: %d deployments\n", len(d.interceptables))
+	fmt.Fprintf(res, "  Intercepts:    ? total, %d local\n", len(d.intercepts))
+	return nil
+}
